try to make the task area workable
Task system was maintaining both a begin and an end date which makes the facility to complicated.  Tasks are basically a reminder system, not a project planning or more complex calendaring feature.

So, changed the logic to work off of 'begin' date only.  If you push a task one week, it pushes the begin date.   Also, sort by 'begin' date as well.